# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smallbiz Marketing Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus];

export const VideoStatus = {
  planning: "planning",
  in_production: "in_production",
  review: "review",
  approved: "approved",
  published: "published",
} as const;

export interface Video {
  id: number;
  title: string;
  description?: string;
  status: VideoStatus;
  scheduledDate?: string | null;
  publishedDate?: string | null;
  folderLink?: string | null;
  platform?: string | null;
  tags?: string | null;
  views?: number | null;
  likes?: number | null;
  comments?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateVideoRequestStatus =
  (typeof CreateVideoRequestStatus)[keyof typeof CreateVideoRequestStatus];

export const CreateVideoRequestStatus = {
  planning: "planning",
  in_production: "in_production",
  review: "review",
  approved: "approved",
  published: "published",
} as const;

export interface CreateVideoRequest {
  title: string;
  description?: string;
  status: CreateVideoRequestStatus;
  scheduledDate?: string | null;
  folderLink?: string | null;
  platform?: string | null;
  tags?: string | null;
}

export type UpdateVideoRequestStatus =
  (typeof UpdateVideoRequestStatus)[keyof typeof UpdateVideoRequestStatus];

export const UpdateVideoRequestStatus = {
  planning: "planning",
  in_production: "in_production",
  review: "review",
  approved: "approved",
  published: "published",
} as const;

export interface UpdateVideoRequest {
  title?: string;
  description?: string;
  status?: UpdateVideoRequestStatus;
  scheduledDate?: string | null;
  publishedDate?: string | null;
  folderLink?: string | null;
  platform?: string | null;
  tags?: string | null;
  views?: number | null;
  likes?: number | null;
  comments?: number | null;
}

export type ShootStatus = (typeof ShootStatus)[keyof typeof ShootStatus];

export const ShootStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Shoot {
  id: number;
  title: string;
  description?: string | null;
  shootDate: string;
  location?: string | null;
  status: ShootStatus;
  videoId?: number | null;
  videoTitle?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateShootRequestStatus =
  (typeof CreateShootRequestStatus)[keyof typeof CreateShootRequestStatus];

export const CreateShootRequestStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface CreateShootRequest {
  title: string;
  description?: string | null;
  shootDate: string;
  location?: string | null;
  status: CreateShootRequestStatus;
  videoId?: number | null;
  notes?: string | null;
}

export type UpdateShootRequestStatus =
  (typeof UpdateShootRequestStatus)[keyof typeof UpdateShootRequestStatus];

export const UpdateShootRequestStatus = {
  scheduled: "scheduled",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateShootRequest {
  title?: string;
  description?: string | null;
  shootDate?: string;
  location?: string | null;
  status?: UpdateShootRequestStatus;
  videoId?: number | null;
  notes?: string | null;
}

export type ApprovalStatus =
  (typeof ApprovalStatus)[keyof typeof ApprovalStatus];

export const ApprovalStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  revision_requested: "revision_requested",
} as const;

export interface Approval {
  id: number;
  videoId: number;
  videoTitle: string;
  status: ApprovalStatus;
  reviewerName?: string | null;
  comments?: string | null;
  submittedAt: string;
  reviewedAt?: string | null;
}

export interface CreateApprovalRequest {
  videoId: number;
  comments?: string | null;
}

export type UpdateApprovalRequestStatus =
  (typeof UpdateApprovalRequestStatus)[keyof typeof UpdateApprovalRequestStatus];

export const UpdateApprovalRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  revision_requested: "revision_requested",
} as const;

export interface UpdateApprovalRequest {
  status: UpdateApprovalRequestStatus;
  reviewerName?: string | null;
  comments?: string | null;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface MonthlyMetric {
  month: string;
  views: number;
  likes: number;
}

export interface Analytics {
  totalVideos: number;
  publishedVideos: number;
  pendingApprovals: number;
  upcomingShoots: number;
  totalViews: number;
  totalLikes: number;
  totalComments: number;
  videosByStatus: StatusCount[];
  viewsByMonth: MonthlyMetric[];
  recentVideos: Video[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};