# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smallbiz Marketing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyMetric } from "./monthlyMetric";
import type { StatusCount } from "./statusCount";
import type { Video } from "./video";

export interface Analytics {
  totalVideos: number;
  publishedVideos: number;
  pendingApprovals: number;
  upcomingShoots: number;
  totalViews: number;
  totalLikes: number;
  totalComments: number;
  videosByStatus: StatusCount[];
  viewsByMonth: MonthlyMetric[];
  recentVideos: Video[];
}