# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smallbiz Marketing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from "./approvalStatus";

export interface Approval {
  id: number;
  videoId: number;
  videoTitle: string;
  status: ApprovalStatus;
  reviewerName?: string | null;
  comments?: string | null;
  submittedAt: string;
  reviewedAt?: string | null;
}