# lib/api-zod/src/generated/types/createApprovalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smallbiz Marketing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApprovalRequest {
  videoId: number;
  comments?: string | null;
}