# lib/api-zod/src/generated/types/shoot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smallbiz Marketing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ShootStatus } from "./shootStatus";

export interface Shoot {
  id: number;
  title: string;
  description?: string | null;
  shootDate: string;
  location?: string | null;
  status: ShootStatus;
  videoId?: number | null;
  videoTitle?: string | null;
  notes?: string | null;
  createdAt: string;
}